metrics: move prometheus sanitizer into metrics
This single-file utility is always used in conjunction with metrics and
doesn't need to live on its own in a single top-level directory.
