+++ src/v/archival/probe.cc
@@ -11,7 +11,7 @@
 #include "archival/probe.h"
 
 #include "config/configuration.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 
 #include <seastar/core/metrics.hh>
 #include <seastar/core/smp.hh>
