+++ src/v/cloud_roles/probe.cc
@@ -11,7 +11,7 @@
 #include "cloud_roles/probe.h"
 
 #include "config/configuration.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 
 #include <seastar/core/metrics.hh>
