+++ src/v/cloud_storage/remote_probe.cc
@@ -12,7 +12,7 @@
 
 #include "cloud_storage/materialized_resources.h"
 #include "metrics/metrics.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 
 #include <seastar/core/metrics.hh>
 #include <seastar/core/smp.hh>
