+++ src/v/cluster/rm_stm.cc
@@ -17,10 +17,10 @@
 #include "cluster/types.h"
 #include "kafka/protocol/wire.h"
 #include "metrics/metrics.h"
+#include "metrics/prometheus_sanitize.h"
 #include "model/fundamental.h"
 #include "model/record.h"
 #include "model/timestamp.h"
-#include "prometheus/prometheus_sanitize.h"
 #include "raft/consensus_utils.h"
 #include "raft/errc.h"
 #include "raft/fundamental.h"
