+++ src/v/cluster/scheduling/leader_balancer_probe.cc
@@ -11,8 +11,8 @@
 #include "cluster/scheduling/leader_balancer_probe.h"
 
 #include "config/configuration.h"
+#include "metrics/prometheus_sanitize.h"
 #include "model/fundamental.h"
-#include "prometheus/prometheus_sanitize.h"
 
 #include <seastar/core/metrics.hh>
