+++ src/v/kafka/group_probe.h
@@ -15,9 +15,9 @@
 #include "kafka/server/member.h"
 #include "kafka/types.h"
 #include "metrics/metrics.h"
+#include "metrics/prometheus_sanitize.h"
 #include "model/fundamental.h"
 #include "model/namespace.h"
-#include "prometheus/prometheus_sanitize.h"
 
 #include <seastar/core/metrics.hh>
