+++ src/v/kafka/latency_probe.h
@@ -13,7 +13,7 @@
 
 #include "config/configuration.h"
 #include "metrics/metrics.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 #include "utils/log_hist.h"
 
 #include <seastar/core/metrics.hh>
