+++ src/v/kafka/server/handlers/handler_probe.cc
@@ -16,7 +16,7 @@
 #include "kafka/protocol/schemata/produce_request.h"
 #include "kafka/server/handlers/handler_interface.h"
 #include "kafka/server/logger.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 
 #include <seastar/core/lowres_clock.hh>
 #include <seastar/core/metrics.hh>
