+++ src/v/kafka/server/quota_manager.cc
@@ -17,7 +17,7 @@
 #include "kafka/server/client_quota_translator.h"
 #include "kafka/server/logger.h"
 #include "metrics/metrics.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 #include "ssx/future-util.h"
 #include "utils/log_hist.h"
