+++ src/v/pandaproxy/probe.cc
@@ -11,7 +11,7 @@
 
 #include "config/configuration.h"
 #include "metrics/metrics.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 #include "ssx/sformat.h"
 
 #include <seastar/core/metrics.hh>
