+++ src/v/raft/group_manager.cc
@@ -12,8 +12,8 @@
 #include "base/likely.h"
 #include "config/configuration.h"
 #include "features/feature_table.h"
+#include "metrics/prometheus_sanitize.h"
 #include "model/metadata.h"
-#include "prometheus/prometheus_sanitize.h"
 #include "raft/group_configuration.h"
 #include "raft/rpc_client_protocol.h"
 #include "resource_mgmt/io_priority.h"
