+++ src/v/raft/probe.cc
@@ -10,8 +10,8 @@
 #include "raft/probe.h"
 
 #include "config/configuration.h"
+#include "metrics/prometheus_sanitize.h"
 #include "model/fundamental.h"
-#include "prometheus/prometheus_sanitize.h"
 
 #include <seastar/core/metrics.hh>
