+++ src/v/security/authorizer.cc
@@ -14,7 +14,7 @@
 #include "acl_store.h"
 #include "config/configuration.h"
 #include "metrics/metrics.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 #include "security/role.h"
 #include "security/role_store.h"
