+++ src/v/storage/kvstore.cc
@@ -13,9 +13,9 @@
 #include "bytes/iobuf.h"
 #include "bytes/iostream.h"
 #include "config/configuration.h"
+#include "metrics/prometheus_sanitize.h"
 #include "model/async_adl_serde.h"
 #include "model/namespace.h"
-#include "prometheus/prometheus_sanitize.h"
 #include "reflection/adl.h"
 #include "ssx/async_algorithm.h"
 #include "storage/parser.h"
