+++ src/v/transform/probe.cc
@@ -11,8 +11,8 @@
 
 #include "probe.h"
 
+#include "metrics/prometheus_sanitize.h"
 #include "model/transform.h"
-#include "prometheus/prometheus_sanitize.h"
 
 #include <seastar/core/metrics.hh>
