+++ src/v/wasm/engine_probe.cc
@@ -13,7 +13,7 @@
 
 #include "base/vassert.h"
 #include "metrics/metrics.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 
 #include <seastar/core/shared_ptr.hh>
 #include <seastar/core/weak_ptr.hh>
