+++ src/v/wasm/transform_probe.cc
@@ -12,7 +12,7 @@
 #include "wasm/transform_probe.h"
 
 #include "config/configuration.h"
-#include "prometheus/prometheus_sanitize.h"
+#include "metrics/prometheus_sanitize.h"
 #include "utils/log_hist.h"
 
 #include <seastar/core/metrics.hh>
